pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ark-discord-bot"
version = "0.1.0"
description = "Discord bot for managing ARK: Survival Ascended server"
authors = [{name = "boxp", email = "boxp@example.com"}]
dependencies = [
    "discord.py==2.3.2",
    "kubernetes==28.1.0",
    "pyyaml==6.0.1",
    "python-dotenv==1.0.0",
    "aiohttp==3.9.1",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "pytest-mock==3.12.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"